requirements.txt
# Core LangChain packages
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0

# LangGraph — required for deep_agent
langgraph>=0.2.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
duckduckgo-search>=6.0.0
